// src/generated/v2/schema/batchCreateScheduledBotResponseDataItem.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Meeting BaaS API
 * API for managing meeting bots, calendar integrations, and webhooks
 * OpenAPI spec version: 2.0.0
 */
import type { BatchCreateScheduledBotResponseDataItemExtra } from "./batchCreateScheduledBotResponseDataItemExtra"

export type BatchCreateScheduledBotResponseDataItem = {
  /**
   * 0-based index of the item in the request array
   * @minimum 0
   * @maximum 9007199254740991
   */
  index: number
  /**
   * The UUID of the created scheduled bot
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  bot_id: string
  /** Custom metadata from the request (null if not provided) */
  extra: BatchCreateScheduledBotResponseDataItemExtra
}